+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Workaround for missing click event in ios safari
+  
 ## [1.2.1] - 2019-09-16
 
 - Replace originalTarget with target to be standards compliant.